move server setup into separate file
